fix(security): harden OAuth token checks and remove insecure deployment defaults
- Harden Apple OAuth verification in auth-service with token format checks, RS256/type enforcement, JWKS key validation, and request timeouts/status checks
- Remove default Grafana admin password from monitoring values and require explicit secure password in deployment scripts
- Require DB_PASSWORD/GRAFANA_ADMIN_PASSWORD in deploy.sh to prevent weak default credentials
- Update env examples and include mobile .env.example for safe local configuration
- Include related repository updates currently staged in working tree

+++ services/health-data-service/src/main/resources/db/migration/V1__create_health_tables.sql
@@ -5,7 +5,7 @@ CREATE TABLE sleep_entries (
     sleep_start TIMESTAMPTZ NOT NULL,
     sleep_end   TIMESTAMPTZ NOT NULL,
     duration_min INT,
-    quality     SMALLINT CHECK (quality BETWEEN 1 AND 5),
+    quality     INTEGER CHECK (quality BETWEEN 1 AND 5),
     notes       TEXT,
     source      VARCHAR(30),
     created_at  TIMESTAMPTZ DEFAULT NOW()
@@ -45,7 +45,7 @@ CREATE TABLE symptom_entries (
     id          UUID PRIMARY KEY DEFAULT gen_random_uuid(),
     user_id     UUID NOT NULL,
     symptom     VARCHAR(100) NOT NULL,
-    intensity   SMALLINT CHECK (intensity BETWEEN 1 AND 10),
+    intensity   INTEGER CHECK (intensity BETWEEN 1 AND 10),
     recorded_at TIMESTAMPTZ NOT NULL,
     notes       TEXT
 );
